requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# HL7 Integration (استخدام إصدارات متوفرة)
hl7==0.4.5
hl7apy==1.3.4

# Async & Background Tasks (اختياري - يمكن تثبيتها لاحقًا)
# celery==5.3.4
# redis==5.0.1
# django-celery-beat==2.5.0
# django-celery-results==2.5.1

# Security & Auth
PyJWT==2.8.0
django-cors-headers==4.3.1
ldap3==2.9.1

# File Handling
Pillow==10.1.0

# Monitoring & Logging
django-extensions==3.2.3

# Documentation
drf-spectacular==0.27.0

# التبعيات الإضافية (اختياري)
# pandas==2.1.4
# numpy==1.26.2
